Add frequency metadata
Adding a new CSV source file, containing 65 new columns:
freq1-freq30: the frequency of each response in the data
per1-per30: the percent of non-system missing values that freqN represents
Only for ordered variables:
obs: Total non-missing observations
min, max: Min and max values
avg: the mean
std: the standard deviation

Removed columns for valueN and labelN where N >30. The new value and label columns represent what values are actually displayed in the data, rather than what was in the value labels. The number 30 was chosen to reflect the total number of values used to distinguish between an ordered variable and a continuous one.

Also added code to handle these new columns, and removed code for the columns no longer in the file.